Statement Factory created
