+++ src/compiler/parser/statement_factory/mod.rs
@@ -1,0 +1,3 @@
+mod statement_factory;
+
+pub use statement_factory::StatementFactory;
